% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pare.down.R
\name{Pare.down}
\alias{Pare.down}
\title{Pare.down}
\usage{
Pare.down(df, Month, Site)
}
\arguments{
\item{df}{your data frame, use the output from the previous function}

\item{Month}{the month you want}

\item{Site}{the site that you want to pare down to}
}
\description{
Pare.down
}
