% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Species.Summary.R
\name{Species.Summary}
\alias{Species.Summary}
\title{Species.Summary}
\usage{
Species.Summary(df, Species, Species1, Abundance)
}
\arguments{
\item{df}{Data produced by Pare.Down function that has been cleaned for a specific month and site}

\item{Species}{the column name}

\item{Species1}{First species that you would like to take out}

\item{Abundance}{Title of column where abundance of species is located}
}
\description{
Species.Summary
}
